checker: Moved the code from main() function to the separate crate
