+++ util/checker/src/cmd/mod.rs
@@ -1,0 +1,5 @@
+mod check_stubs_compile;
+
+pub fn check_stubs_compile() {
+    check_stubs_compile::check_stubs_compile();
+}
